start writing hwregtrace
