+++ RiscSimulator/IO.h
@@ -11,3 +11,4 @@ extern int16_t irqreturn;
 extern char timerenable; // 1 bit
 extern int32_t timercurrent;
 extern int32_t timermax;
+extern char *hwregtraceFileName;
