+++ RiscSimulator/RiscSimulator.vcxproj.filters
@@ -29,13 +29,13 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="singltons.h">
-      <Filter>Source Files</Filter>
+    <ClInclude Include="IO.h">
+      <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="opcodes.h">
-      <Filter>Source Files</Filter>
+      <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="IO.h">
+    <ClInclude Include="singltons.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
